conditional build for linux and others
